+++ src/plugins/utils/utils-rxdb-version.ts
@@ -1,4 +1,4 @@
 /**
  * This file is replaced in the 'npm run build:version' script.
  */
-export const RXDB_VERSION = '15.5.0';
+export const RXDB_VERSION = '15.6.0';