+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 # RxDB Changelog
 
 <!-- CHANGELOG NEWEST -->
-
+- FIX firestore replication not working with schema validation [#5572](https://github.com/pubkey/rxdb/issues/5572)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->